Uploaded all required code [FINAL]
These files no longer include datasets for which I'm unsure about my right to publish publicly, yet it does include the webscrapers useful for gathering the required data by yourself. The code has been just a tool for testing my hypotheses, so I'm aware of the ugliness of some of it.